docker_streamlit_app/requirements.txt
streamlit
pandas
matplotlib.pyplot
math
seaborn